Move leaf node to own module
**Motivation**
The `representation.rs` file is pretty large and it is difficult to read
at times, this change makes that slightly better.
